finished updating tests
